backend/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
boto3==1.20.26
botocore==1.23.54
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cliapp==1.0.9
click==8.1.7
cmdtest==0.1.1
comm==0.2.2
contourpy==1.2.1
crispy-bootstrap4==2024.1
cryptography==41.0.7
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
distro-info==1.0
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==4.2.7
django-allauth==0.63.6
django-anymail==10.2
django-ckeditor-5==0.2.17
django-cors-headers==3.14.0
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.2
django-jazzmin==2.6.0
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
environs==10.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.52.4
fpdf==1.7.2
fqdn==1.5.1
FreeSimpleGUI==5.0.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.6
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==4.6.4
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
moviepy==1.0.3
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
netifaces==0.11.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.0
openai==1.34.0
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.6.0
pyOpenSSL==24.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pypng==0.20220715.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
pytest>=8.0.0,<9.0.0
pytest-django==4.11.1 # Added for Django integration with Pytest
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
s3transfer==0.5.2
SecretStorage==3.3.1
Send2Trash==1.8.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.34.0
stripe==7.9.0
systemd-python==234
tenacity==8.3.0
terminado==0.18.1
tinycss2==1.3.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
wadllib==1.3.6
watchdog==4.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
whitenoise==6.7.0
zipp==1.0.0